Updated loadContent, todo and refactored code
What changes you have made?
- loadContent
- Now loading content from stdin also.
- todo.txt
- Marked all the things I did.
- Refactored code
- Make function async
- Package.json
- Added webtest script

Why this change is necessary?
-loadContent
- As per the second requirement we should read ointent from stdin
when no files are given.
- Refactored code
- Earlier as everthing is sync it won't wait for the standard
input. It opens the read stream and do all the after processes. Instead
we want to wait for the content then perfrom those operations.
That's why we make is async
- Package.json
- So, that we can the coverage in web view.

What is the impact of the change?
- loadContent
- Now we can take user input via command line
- Package.json
- Now, we can check what all the lines which are not covered.